Module BantorraBasis

This library implements basic routines used by other libraries.

module File : sig ... end

Routines to handle file paths and basic I/O.

module Marshal : sig ... end

Routines for serialization.

module Xdg : sig ... end

Routines to calculate the directories in the XDG standard while having reasonable default values on major platforms.

module Util : sig ... end

Other routines that are difficult to classify.