Module BantorraBasis__.Marshal

Types

type value = Ezjsonm.value

The type suitable for marshalling. This is the universal type to exchange information within the framework.

Serialization

val of_json : string -> (value, [> `FormatError of string ]) Stdlib.result

A function that deserializes a value.

val read_json : string -> (value, [> `FormatError of string | `SystemError of string ]) Stdlib.result

read_json path v reads and deserializes the content of the file at path.

val to_json : ?⁠minify:bool -> value -> string

A function that serializes a value.

val write_json : ?⁠minify:bool -> string -> value -> (unit, [> `FormatError of string | `SystemError of string ]) Stdlib.result

unsafe_write_json path v writes the serialization of v into the file at path.

Helper Functions

val of_string : string -> value

Embedding a string into a value.

val to_string : value -> (string, [> `FormatError of string ]) Stdlib.result

Projecting a string out of a value.

val of_ostring : string option -> value

Embedding an optional string into a value.

val to_ostring : value -> (string option, [> `FormatError of string ]) Stdlib.result

Projecting an optional string out of a value.

val of_list : ('a -> value) -> 'a list -> value

Embedding a list into a value.

val to_list : (value -> ('a, [> `FormatError of string ] as 'e) Stdlib.result) -> value -> ('a list, 'e) Stdlib.result

Projecting a list out of a value.

val of_olist : ('a -> value) -> 'a list option -> value

Embedding an optional list into a value.

val to_olist : (value -> ('a, [> `FormatError of string ] as 'e) Stdlib.result) -> value -> ('a list option, 'e) Stdlib.result

Projecting an optional list out of a value.

val parse_object : ?⁠required:string list -> ?⁠optional:string list -> value -> ((string * value) list * (string * value) list, [> `FormatError of string ]) Stdlib.result

Projecting an associative list out of a value.

parameter required

Names of required fields. By default, it is [].

parameter optional

Names of optional fields. By default, it is [].

returns

A pair of an associative list for required fields and that for optional fields. Missing optional fields will be associated with `Null. In other words, a missing "opt" field in a JSON object is treated as "opt": null.

val parse_object_or_null : ?⁠required:string list -> ?⁠optional:string list -> value -> (((string * value) list * (string * value) list) option, [> `FormatError of string ]) Stdlib.result

Projecting an optional associative list out of a value. See parse_object.

val dump : Stdlib.Format.formatter -> value -> unit

An ugly-printer for value.