Module BantorraCache.Database

type t

type of database handles

val init : root:string -> t

init ~root initialize the database rooted at root.

parameter root

The root directory of the database. If not existing, it will be created.

val save_state : t -> unit

save_state db saves the state of the database back to the disk.

val digest_of_item : key:BantorraBasis.Marshal.value -> value:BantorraBasis.Marshal.t -> Stdlib.Digest.t

val replace_item : t -> key:BantorraBasis.Marshal.value -> value:BantorraBasis.Marshal.t -> Stdlib.Digest.t

replace_item db ~key ~value saves the content on disk and return a digest that can be used in find_item_opt. It overwrites the content indexed by the same key.

val find_item_opt : t -> key:BantorraBasis.Marshal.value -> digest:Stdlib.Digest.t option -> BantorraBasis.Marshal.t option

find_item_opt db ~key ~digest:(Some digest) tries to retrive the cached result indexed by the key. The content will be checked against the provided digest. find_opt db ~key ~digest:None is the same except that it skips the digest checking.

returns

The function returns None if there is no applicable cache or there is an error during decoding.

returns

Some j The cached result is j.