Module BantorraCache__.Store

Types

type t

The type of cache stores.

Initialization

val init : root:string -> t

init ~root initialize the store rooted at root. It assumes it has exclusive control over the files and directories under root.

parameter root

The root directory of the cache store. It is assumed that the directory already exists.

val save_state : t -> unit

save_state s saves the current state of the cache back to the disk.

Accessors

val replace_item : t -> key:BantorraBasis.Marshal.value -> value:BantorraBasis.Marshal.t -> Stdlib.Digest.t

replace_item s ~key ~value saves the content to disk and return a digest that can be used in find_item_opt. It overwrites the content indexed by the same key.

val digest_of_item : key:BantorraBasis.Marshal.value -> value:BantorraBasis.Marshal.t -> Stdlib.Digest.t

digest_of_item ~key ~value calculates the same digest as if replace_item would have returned without touching the store. This is useful when the cache is disabled but the digest is still needed. (Currently, the option to disable cache is not implemented.)

val find_item_opt : t -> key:BantorraBasis.Marshal.value -> digest:Stdlib.Digest.t option -> BantorraBasis.Marshal.t option

find_item_opt s ~key ~digest:(Some digest) tries to retrieve the cached result indexed by the key. The content will be checked against the provided digest. find_opt db ~key ~digest:None is the same except that it skips the digest checking.

returns

None if there is no applicable cache or there is an error during decoding.

returns

Some j if the cached result is j.