Module BantorraResolvers.Waypoint

Resolver based on local waypoints.

Argument Format

The resolver accepts simple JSON strings as library names.

Waypoints and Landmark Files

A waypoint is either direct or indirect. A direct waypoint points to the library root in the file system directly. Its JSON representation is as follows:

{
  "cool": {
    "at": ["mylib", "cool"]
  }
}

This means the root of the cool library is at the file system path mylib/cool. An indect waypoint is pointing to another waypoint, possibly with a new library name to look up. Its JSON representation is:

{
  "cool": {
    "next": ["mylib"]
  }
}

This means one should look for the landmark file under the mylib directory for further instructions. One can also specify the rename field in case a new library name should be used:

{
  "cool.basis": {
    "next": ["cool"],
    "rename": "basis"
  }
}

A complete landmark file looks like this:

{
  "format": "1.0.0",
  "waypoints": {
    "cool.basis": {
      "next": ["path", "to", "another", "directory"],
      "rename": "cool.basis.new"
    },
    "hello": {
      "at": ["vendor", "lib", "hello"]
    }
  }
}

There are two phases in locating the library using waypoints. In the first phase, the resolver starts from the root of the library that imports the target library, going up in the file system tree until it finds a landmark file with an applicable waypoint (that is, with a matching field in the waypoints JSON object). The resolution fails if no much landmark files are found. If such a landmark file is found, then the resolver enters the second phase. During the second phose, it follow the waypoints until the root of the target library is found. The difference from the first phase is that if the library the resolver is looking for is not explicitly listed in the waypoints object, the resolution fails immediately. In comparison, the resolver would climb up the file system tree during the first phase. This is to prevent unintentional infinite looping due to simple typos.

The Builder

val resolver : ?⁠eager_resolution:bool -> landmark:string -> Bantorra.Resolver.t

resolver ~eager_resolution ~landmark construct a resolver for the specified file name landmark.

parameter eager_resolution

Whether full resolution is performed to check the validity of library names. If the value is false, the resolver will only check whether the argument to the resolver is well-formed. The default value is false.

parameter landmark

The name of the special landmark files that the resolver should look for.

val clear_cached_landmarks : unit -> unit

Landmark files are all cached to reduce I/O load, but perhaps you are writing some tool to modify landmark files. In this case, one should call this function to reload all landmark files.