Module BantorraResolvers__.OCamlFind

Introduction

This is a resolver similar to the fixed-table resolver FixedTable except that each path is relative the package directory given by the OCaml library findlib. The package directory is used by the ocamlfind tool as part of the standard OCaml toolchain.

The resolver accepts simple YAML strings as library names.

The Builder

val resolver : package_name:string -> dict:(string * string) list -> Bantorra.Resolver.t

resolver ~package_name ~dict constructs a resolver based on the package directory given by findlib and the mapping dict. All paths are normalized and turned into absolute paths with respect to the current working directory using BantorraBasis.File.normalize_dir.