Module BantorraResolvers__.Waypoint

Introduction

The waypoint resolver walks around the file system to resolve library names. It tries to find special files (called landmarks) that contain waypoints, which help the resolver to find the roots of target libraries. All landmark files have the same file name specified by the argument given to resolver.

The resolver accepts simple YAML strings as library names.

A waypoint is either direct or indirect. A direct waypoint points to the library root directly. Its YAML representation is as follows:

cool:
  at: [mylib, cool]

This means the root of the cool library is at mylib/cool. An indect waypoint is pointing to another waypoint, possibly with a new library name to look up. Its YAML representation is:

cool:
  next: [mylib]

This means one should look for the landmark file under the mylib directory for further instructions. One can also specify the rename field in case a new library name should be used:

cool.basis:
  next: [cool]
  rename: basis

A complete landmark file looks like this:

format: "1.0.0"
waypoints:
  cool.basis:
    next: [cool]
    rename: basis
  hello:
    at: [misc, hello]

The resolver starts from the root of the library that imports the target library, going up in the file system tree until it finds a landmark file with an applicable waypoint. It will then follow the waypoints until the root of the target library is found. Note that the resolver only climbs up the file system tree once, at the beginning. The resolution immediately fails if an indirect waypoint points to a landmark file with no applicable waypoint. This would prevent unintentional infinite looping due to simple typos.

The Builder

val resolver : strict_checking:bool -> landmark:string -> Bantorra.Resolver.t

resolver ~strict_checking ~landmark construct a resolver for the specified file name landmark.

parameter strict_checking

Whether full resolution is performed to check the validity of library names. If the value is false, the resolver will only check whether all dependencies are well-formed when loading a new library.

parameter landmark

The name of the special landmark files that the resolver should look for.

val clear_cached_landmarks : unit -> unit

Landmark files are all cached to reduce I/O load, but perhaps you are writing some tool to modify landmark files. In this case, one should call this function to reload all landmark files.