Module BantorraRouters.FixedTable

A router uses a fixed mapping table.

Argument Format

The router accepts simple JSON strings as library names.

Builder

val router : dict:(string * BantorraBasis.File.filepath) list -> (Bantorra.Router.t, [> `InvalidRouter of string ]) Stdlib.result

router ~dict construct a router based on the mapping dict. All paths in dict will be normalized and turned into absolute paths with respect to the current working directory.