Module BantorraRouters.OCamlFind

A router using findlib.

Argument Format

The router accepts JSON strings as library names.

Builder

val router : package_name:string -> dict:(string * string) list -> (Bantorra.Router.t, [> `InvalidRouter of string ]) Stdlib.result

router ~package_name ~dict constructs a router based on the package directory given by findlib and the mapping dict. All paths are normalized and turned into absolute paths with respect to the current working directory using BantorraBasis.File.normalize_dir.