Module BantorraRouters__Direct

A router that directly accepts the library root.

Argument Format

The router accepts a string as an absolute path or a path relative to the library root.

Builder

val router : Bantorra.Router.t