Module Bantorra__.Library

Types

type unitpath = Bantorra.Anchor.unitpath

The type of unit paths.

type t

The type of libraries.

Initialization

val init : anchor:string -> root:BantorraBasis.File.filepath -> t

Initite a library rooted at root where the name of the anchor file is anchor.

val locate_anchor_from_file : anchor:string -> suffix:string -> BantorraBasis.File.filepath -> BantorraBasis.File.filepath * unitpath

See Manager.locate_anchor_from_file.

val locate_anchor_from_dir : anchor:string -> BantorraBasis.File.filepath -> BantorraBasis.File.filepath * unitpath

See Manager.locate_anchor_from_dir.

val locate_anchor_from_cwd : anchor:string -> BantorraBasis.File.filepath * unitpath

See Manager.locate_anchor_from_cwd.

Accessor

val iter_deps : (Bantorra.Anchor.lib_ref -> unit) -> t -> unit

Iterate over all dependencies listed in the anchor.

Hook for Library Managers

val to_unitpath : global:(current_root:BantorraBasis.File.filepath -> Bantorra.Anchor.lib_ref -> unitpath -> t * unitpath) -> t -> unitpath -> t * unitpath

to_unitpath ~global lib unitpath resolves unitpath and returns the eventual library where the unit belongs and the local unit path pointing to the unit.

parameter global

The global resolver for unit paths pointing to other libraries.

val to_filepath : global:(current_root:BantorraBasis.File.filepath -> Bantorra.Anchor.lib_ref -> unitpath -> suffix:string -> t * BantorraBasis.File.filepath) -> t -> unitpath -> suffix:string -> t * BantorraBasis.File.filepath

to_filepath ~global lib unitpath ~suffix resolves unitpath and returns the eventual library where the unit belongs and the underlying file path of the unit. It is similar to to_unitpath but returns a file path instead of a unit path.

parameter global

The global resolver for unit paths pointing to other libraries.

parameter suffix

The suffix shared by all the units in the file system.