Module Bantorra__Library

Types

type unitpath = Bantorra__.Anchor.unitpath

The type of unit paths.

type t

The type of libraries.

Initialization

val load_from_root : find_cache:(string -> t option) -> anchor:string -> BantorraBasis.File.filepath -> (t, [> `InvalidLibrary of string ]) Stdlib.result

val load_from_dir : find_cache:(string -> t option) -> anchor:string -> BantorraBasis.File.filepath -> (t * unitpath option, [> `InvalidLibrary of string ]) Stdlib.result

val load_from_unit : find_cache:(string -> t option) -> anchor:string -> suffix:string -> BantorraBasis.File.filepath -> (t * unitpath option, [> `InvalidLibrary of string ]) Stdlib.result

Accessor

val root : t -> BantorraBasis.File.filepath

val iter_routes : (router:string -> router_argument:BantorraBasis.Marshal.value -> (unit, 'e) Stdlib.result) -> t -> (unit, 'e) Stdlib.result

Hook for Library Managers

val resolve : global:(starting_dir:BantorraBasis.File.filepath -> router:string -> router_argument:BantorraBasis.Marshal.value -> unitpath -> suffix:string -> (t * unitpath * BantorraBasis.File.filepath, [> `UnitNotFound of string ] as 'e) Stdlib.result) -> t -> unitpath -> suffix:string -> (t * unitpath * BantorraBasis.File.filepath, 'e) Stdlib.result

to_unitpath ~global lib unitpath resolves unitpath and returns the eventual library where the unit belongs and the local unit path pointing to the unit.

parameter global

The global resolver for unit paths pointing to other libraries.