Module Cache.Storage

type json_value = Ezjsonm.value

The type of JSON values. Please consult the JSON standard and the OCaml package jsonm for limitations.

This is intended to be compatible with the OCaml package ezjsonm.

type json = Ezjsonm.t

The type of JSON documents. The top-level must be an array or an object.

This is intended to be fully compatible with the OCaml package ezjsonm.

type t

val init : root:string -> t

val replace : t -> key:json -> value:json -> Stdlib.Digest.t

val find_opt : t -> key:json -> digest:Stdlib.Digest.t option -> json option