Module Cache__Storage

type json_value = [

| `Null

| `Bool of bool

| `Float of float

| `String of string

| `A of json_value list

| `O of (string * json_value) list

]

The type of JSON values. Please consult the JSON standard and the OCaml package jsonm for limitations.

This is intended to be compatible with the OCaml package ezjsonm.

type json = [

| `A of json_value list

| `O of (string * json_value) list

]

The type of JSON documents. The top-level must be an array or an object.

This is intended to be fully compatible with the OCaml package ezjsonm.

module Database : sig ... end