Module BantorraResolvers__UserConfig

Argument Format

The resolver takes a JSON argument in one of the following formats:

{
  "name": "bantorra",
  "version": "0.1.0"
}

{
  "name": "bantorra",
  "version": null
}

{ "name": "bantorra" }

A missing version is understood as the version null. Therefore, the last two specifications are identical. Versions are compared using structural equality. There is no smart comparison, version globbing, or ordering between versions. Each version is completely independent of each other; the version null only matches null, not any string version. Not even the string "null".

However, one could have multiple versions of the same library registered in the user configuration file, as long as the versions in use all point to the same library on disk. (For example, one may have both versions "1.0" and "1" pointing to /usr/lib/hello/1.0, which means they can be used interchangeably.) Any attempt to load incompatible versions during the program execution would abort the library resolution.

Configeration Format

Builder

val resolver : app_name:string -> config:string -> Bantorra.Resolver.t

resolver ~app_name ~config constructs a resolver that reads the user configuration. The location of the user configuration is given by BantorraBasis.Xdg.get_config_home concatenated with config. All paths are normalized and turned into absolute paths with respect to the current working directory using BantorraBasis.File.normalize_dir .

If the configuration file does not exist, an empty mapping is used, which means the resolver would reject every request.

parameter app_name

The application name for generating a suitable directory to put the configuration file.

parameter config

The file name of the configuration file.

val clear_cached_configs : unit -> unit

Configuration files are all cached to reduce I/O load, but perhaps you are bypassing this module to modify them. In that case, one should call this function to force rereading configuration files.

Configuration I/O

type versioned_library = {

name : string;

version : string option;

}

The type of versioned library names. None corresponds to null and Some ver corresponds to explicit versions.

type filepath = string

type config

The type of configurations as association lists.

val default_config : config

Default configuration that is empty.

val read : app_name:string -> config:filepath -> config

Try to read the configuration file. Note that the results are cached. See clear_cached_configs. If the configuration file does not exist or is ill-formated, then the default configuration (the empty mapping) is returned. The cache will be updated accordingly.

parameter app_name

The application name for generating a suitable directory to put the configuration file.

parameter config

The file path of the configuration file.

val lookup : name:string -> version:string option -> config -> filepath option

val unsafe_write : app_name:string -> config:string -> config -> unit

Write the configuration file. Do not use this function unless you know what you are doing. Due to some questionable design of the underlying OCaml YAML package yaml, strings such as "1.0" and "null" will not be serialized correctly. It is thus recommended to edit the configuration files directly before the package yaml is either fixed or replaced. See the issue on GitHub.

The cache within this module will be updated upon successful writing. See clear_cached_configs. However, the caveat in the package yaml means the written configuration file might be broken or different upon rereading.

parameter app_name

The application name for generating a suitable directory to put the configuration file.

parameter config

The file name of the configuration file.