Module Bantorra__Anchor

Types

type t

The type of anchors.

type unitpath = string list

The type of unit paths.

Anchor I/O

val read : string -> (t, [> `FormatError of string | `SystemError of string ]) Stdlib.result

read path read the content of an anchor file.

Accessors

val iter_routes : (router:string -> router_argument:BantorraBasis.Marshal.value -> (unit, 'e) Stdlib.result) -> t -> (unit, 'e) Stdlib.result

val dispatch_path : t -> unitpath -> (string * BantorraBasis.Marshal.value * unitpath) option

dispatch_path a p routes the unit path p to Some (ref, p') if it points to a unit in another library referenced by ref and p', or None if it is a local unit path. The dispatching is done by longest prefix match.

val path_is_local : t -> unitpath -> bool